make sidebar menu scroll vertically
