--- public/js/header.js
@@ -1,14 +1,0 @@
-const header = {
-  init() {
-    this.bindCloseAnywhere();
-  },
-
-  bindCloseAnywhere() {
-    const bgClickElement = document.querySelector('.js-sidebar-menu');
-    bgClickElement.addEventListener('click', () => {
-      location.href = '#';
-    });
-  },
-};
-
-export default header;
